// packages/maker/src/api/common/OntologyEntityTypeEnum.ts
/*
 * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export enum OntologyEntityTypeEnum {
  OBJECT_TYPE = "OBJECT_TYPE",
  LINK_TYPE = "LINK_TYPE",
  INTERFACE_TYPE = "INTERFACE_TYPE",
  SHARED_PROPERTY_TYPE = "SHARED_PROPERTY_TYPE",
  ACTION_TYPE = "ACTION_TYPE",
  VALUE_TYPE = "VALUE_TYPE",
}